fix: address pr review findings (round 1)
- split redemption-codes-page.tsx into MintSection + HistorySection siblings
  so the file drops the oxlint-disable max-lines-per-function suppression
- add integration tests for /_/redemption-codes (feature-gate, mint submit,
  history fetch, cache invalidation on mint)
- add EXTRA_STAFF_USER_IDS env-override tests for GET (list) to mirror the
  existing POST (mint) coverage
- wrap recordRedeemAttempt inside the redeem catch with a nested try/catch
  so a secondary accounting-write failure never swallows the original
  redeem error
